@@ -36,10 +36,14 @@
   </scm>
 
   <properties>
-    <renjin.version>RELEASE</renjin.version>
-    <renjin.version.for.plugin>3.5-beta76</renjin.version.for.plugin>
-    <maven.compiler.source>1.8</maven.compiler.source>
-    <maven.compiler.target>1.8</maven.compiler.target>
+    <!--renjin.version>RELEASE</renjin.version-->
+    <!--renjin.version.for.plugin>3.5-beta76</renjin.version.for.plugin-->
+    <renjin.version>3.5-dev</renjin.version>
+    <renjin.version.for.plugin>3.5-dev</renjin.version.for.plugin>
+    <javafx.version>17.0.2</javafx.version>
+    <java.version>11</java.version>
+    <maven.compiler.source>11</maven.compiler.source>
+    <maven.compiler.target>11</maven.compiler.target>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
   </properties>
